+++ src/pages/Search/components/CurrentRegionButton/index.tsx
@@ -1,6 +1,6 @@
 import CustomButton from '@/components/CustomButton';
 import LocationIcon from '@/components/icon/Location';
-import { S } from './RegionSetButton.style';
+import { S } from './style';
 import useAppSelector from '@/hooks/useAppSelector';
 import useAppDispatch from '@/hooks/useAppDispatch';
 import { currentRegionActions } from '@/redux/slice/currentRegionSlice';
@@ -8,7 +8,7 @@ import { useNavigate } from 'react-router-dom';
 import { MY_REGION } from '@/constants/localStorage/key';
 import useGeolocation from '@/hooks/useGeolocation';
 
-const RegionSetButton = () => {
+const CurrentRegionButton = () => {
   const currentRegion = useAppSelector((state) => state.currentRegion.value);
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
@@ -46,4 +46,4 @@ const RegionSetButton = () => {
   );
 };
 
-export default RegionSetButton;
+export default CurrentRegionButton;
