+++ src/pages/Search/components/Index/RegionItem/style.ts
