+++ src/pages/Search/components/RecentSearchList/style.ts
