+++ src/pages/Search/components/SearchHeader/style.ts
